requirements.txt
fastapi~=0.115.4
qrcode~=8.0
pydantic~=2.9.2
PyJWT~=2.9.0
passlib[bcrypt]

SQLAlchemy~=2.0.36
bcrypt~=4.2.0